Make Table initiable with container and task context
Samza table implementations might have the need to access global config
or data structures in the container/task context, eg. registering metrics
with MetricsRegistry. This change adds a init() method to ReadableTable
similar to other initable entities in Samza. TableProvider will invoke
this init before handing out Table instances to the operators.

Also made TableProvider initable which is invoked by TableManager.
With this, we do not need a special init for local table specifically.
This is possible because we can obtain the storage engines directly
from TaskContext.